fix: resolve Path Injection vulnerability in file handling

+++ artists-heaven-backend/src/main/java/com/artists_heaven/product/ProductService.java
@@ -1,5 +1,6 @@
 package com.artists_heaven.product;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -17,6 +18,8 @@ import org.springframework.web.multipart.MultipartFile;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private static final String UPLOAD_DIR = "artists-heaven-backend/src/main/resources/product_media/";
+    private static final Path TARGET_PATH = new File(UPLOAD_DIR).toPath().normalize();
 
     public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
@@ -44,12 +47,15 @@ public class ProductService {
 
     public List<String> saveImages(List<MultipartFile> images) {
         List<String> imageUrls = new ArrayList<>();
-        String uploadDir = "artists-heaven-backend/src/main/resources/product_media"; // Define tu ruta de
-                                                                                      // almacenamiento
 
         for (MultipartFile image : images) {
             String fileName = StringUtils.cleanPath(image.getOriginalFilename());
-            Path targetPath = Paths.get(uploadDir, fileName);
+            Path targetPath = Paths.get(UPLOAD_DIR, fileName).normalize();
+
+            if (!targetPath.startsWith(TARGET_PATH)) {
+                throw new IllegalArgumentException("Entry is outside of the target directory");
+            }
+
             try {
                 // Guardar la imagen en el directorio
                 Files.copy(image.getInputStream(), targetPath);
@@ -65,11 +71,16 @@ public class ProductService {
     }
 
     public List<String> deleteImages(List<MultipartFile> removedImages) {
-        String uploadDir = "artists-heaven-backend/src/main/resources/product_media";
         List<String> imagesToDelete = new ArrayList<>();
+
         for (MultipartFile image : removedImages) {
             String fileName = StringUtils.cleanPath(image.getOriginalFilename());
-            Path targetPath = Paths.get(uploadDir, fileName);
+            Path targetPath = Paths.get(UPLOAD_DIR, fileName).normalize();
+
+            if (!targetPath.startsWith(TARGET_PATH)) {
+                throw new IllegalArgumentException("Entry is outside of the target directory");
+            }
+
             try {
                 Files.delete(targetPath);
             } catch (IOException e) {
@@ -100,35 +111,34 @@ public class ProductService {
     public void updateProduct(Product product, List<MultipartFile> removedImages, List<MultipartFile> newImages, ProductDTO productDTO) {
         // Manejo de imágenes
         List<String> existingImages = product.getImages();
-    
+
         if (removedImages != null && !removedImages.isEmpty()) {
             // Eliminar imágenes del servidor
             List<String> removedImagesCast = deleteImages(removedImages);
-    
+
             // Filtrar las imágenes existentes para eliminar las indicadas
             existingImages.removeIf(removedImagesCast::contains);
         }
-    
+
         if (newImages != null && !newImages.isEmpty()) {
             // Guardar nuevas imágenes en el servidor y obtener sus URLs
             List<String> newImageUrls = saveImages(newImages);
-    
+
             // Combinar las imágenes existentes con las nuevas
             existingImages.addAll(newImageUrls);
         }
-    
+
         // Actualizar el producto con las imágenes finales
         productDTO.setImages(existingImages);
-    
+
         // Actualización de los demás datos del producto
         product.setName(productDTO.getName());
         product.setDescription(productDTO.getDescription());
         product.setPrice(productDTO.getPrice());
         product.setSize(productDTO.getSizes());
         product.setCategories(productDTO.getCategories());
-    
+
         // Guardar el producto actualizado
         save(product);
-
     }
 }
